+++ ChangeLog.md
@@ -6,9 +6,12 @@
 
 1. AzCopy now grabs proxy details (sans authentication) from the Windows Registry using `mattn/go-ieproxy`.
 
-### Bug fixes
+### Bug fix
 
+1. Added error to using Azure Files without a SAS token (invalid auth configuration)
 1. AzCopy v10 now outputs a sensible error & warning when attempting to authenticate a storage account business-to-business
+1. `--log-level=none` now drops no logs, and has a listing in `--help`
+1. Fixed bug where piping was not picking up the service version override, making it not work well against Azure Stack.
 
 ## Version 10.1.2
 
@@ -127,4 +130,3 @@ information, including those needed to set the new headers.
    1. excludedBlobType -> excluded-blob-type
    1. outputRaw (in "list" command) -> output
    1. stdIn-enable (reserved for internal use) -> stdin-enable
-